Correct ConstantValue's hanlding of sign extension vs padding
Fixes #18188

The bug arose because of an interaction of this line (new in VS2017)

``` c#
                _builder.EmitIntConstant(endConstant.Int32Value - startConstant.Int32Value);
```

in `SwitchIntegralJumpTableEmitter` and an incorrect implementation of `Int32Value` for constants of type ushort and ubyte. Specifically, the implementation causes improper sign extension of what should be treated as an unsigned value.

We never caught this before because we didn't rely on it. Constant folding in the compiler goes through another (redundant) code path that gets the padding vs sign extension correct. It is possible that, due to this fix, we can simplify the implementation of some constant folding code in binding. I don't do that here, as I want to keep this bug fix as simple and risk-free as possible so it can be included in the next available update.

I don't have a test case for the fix to `Int64Value`. It is possible that there are no symptoms of that instance of the bug. However, I'd rather fix the bug (even if we don't have a concrete symptom where it can occur) than retain the risk of another episode like this bug.